+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - add ``once`` param to ``String#replace`` and ``Regex#replace`` methods
     - now only replace first occurrence of pattern
 - add ``start``, ``end`` method to ``RegexMatch`` type. now get offsets of match string
+- improve error message of builtin ``shctl`` command. now show subcommand name
 
 #### LSP